03-telegram-bot-gemini/requirements.txt
google-generativeai
python-telegram-bot
sqlalchemy
psycopg2-binary
pandas
tabulate
python-dotenv
